backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pandas>=2.2.0
numpy>=2.1.0
scikit-learn>=1.4.0
# TensorFlow for Python 3.12
tensorflow>=2.15.0
# Keras is included with TensorFlow
matplotlib>=3.8.0
openai==1.2.3
python-dotenv==1.0.0
pydantic>=2.6.0
pymongo==4.6.1
pytest==7.4.3
pytest-order==1.2.0
httpx==0.25.0
pytest-cov==4.1.0
pytest-asyncio==0.21.1 